students/k3340/Yakshin_Artemiy/Lr_1/practice_1_3/requirements.txt
fastapi[all]
sqlmodel
psycopg2-binary
alembic
python-dotenv
typing_extensions
